requirements.txt
Django==2.2
django-countries==7.1
django-crispy-forms==1.9.2
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
